feat(react-playground): move url modification code block to react sdk
