+++ packages/react-playground/src/pages/PrivatePage/index.tsx
@@ -1,20 +1,8 @@
-import queryString from 'query-string';
-import React, { useEffect } from 'react';
+import React from 'react';
 
 import styles from './index.module.scss';
 
 const PrivatePage = () => {
-  useEffect(() => {
-    if (window.location.href.includes('?')) {
-      const { url: baseUrl, query: queryObject } = queryString.parseUrl(window.location.href);
-      if (baseUrl && Object.keys(queryObject).length > 0) {
-        const { code, state, ...filteredEntries } = queryObject;
-        const newUrl = queryString.stringifyUrl({ url: baseUrl, query: filteredEntries });
-        history.replaceState(history.state, '', newUrl);
-      }
-    }
-  }, []);
-
   return (
     <div className={styles.wrapper}>
       <div className={styles.title}>Private Page</div>
