+++ packages/react/src/provider.tsx
@@ -1,4 +1,5 @@
 import LogtoClient, { ConfigParameters } from '@logto/client';
+import queryString from 'query-string';
 import React, { useCallback, useEffect, useReducer, useState } from 'react';
 
 import { defaultAuthState } from './auth-state';
@@ -72,6 +73,14 @@ export const LogtoProvider = ({ logtoConfig, children }: LogtoProviderProperties
       try {
         await logtoClient.handleCallback(uri);
         dispatch({ type: 'HANDLE_CALLBACK_SUCCESS', payload: { claims: logtoClient.getClaims() } });
+        if (window.location.href.includes('?')) {
+          const { url: baseUrl, query: queryObject } = queryString.parseUrl(window.location.href);
+          if (baseUrl && Object.keys(queryObject).length > 0) {
+            const { code, state, ...filteredEntries } = queryObject;
+            const newUrl = queryString.stringifyUrl({ url: baseUrl, query: filteredEntries });
+            history.replaceState(history.state, '', newUrl);
+          }
+        }
       } catch (error: unknown) {
         dispatch({ type: 'ERROR', payload: { error } });
       }
